FEAT: 코드 블럭 이름 수정 다이어로그 기능 및 디자인 수정
- 다이어로그 위젯을 별도 파일로 분리하고, 다이어로그 폴더로 묶어서 관리
- 다이어로그 창이 세로로 길게 늘어지던 디자인 수정
- 코드블럭 이름 수정시 이름을 쉽게 수정할 수 있게 selection 전체 지정
