+++ lib/page/sheet_viewer/sheet_viewer.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:chord_everdu/data_class/sheet_info.dart';
 import 'package:chord_everdu/page/sheet_viewer/widget/ChordBlock.dart';
 import 'package:chord_everdu/page/sheet_viewer/widget/chord_keyboard/chord_keyboard.dart';
-import 'package:chord_everdu/page/sheet_viewer/widget/edit_sheet_dialog.dart';
+import 'package:chord_everdu/page/sheet_viewer/widget/dialog/edit_sheet_dialog.dart';
 import 'package:chord_everdu/page/sheet_viewer/widget/new_chord_block_button.dart';
-import 'package:chord_everdu/page/sheet_viewer/widget/sheet_report_dialog.dart';
+import 'package:chord_everdu/page/sheet_viewer/widget/dialog/sheet_report_dialog.dart';
 import 'package:chord_everdu/page/sheet_viewer/widget/sheet_viewer_control_bar.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
